+++ src/compiler/types/generate-prop-types.ts
@@ -1,6 +1,11 @@
 import type * as d from '../../declarations';
 import { getTextDocs } from '@utils';
 
+/**
+ * Generates the individual event types for all @Prop() decorated events in a component
+ * @param cmpMeta component runtime metadata for a single component
+ * @returns the generated type metadata
+ */
 export const generatePropTypes = (cmpMeta: d.ComponentCompilerMeta): d.TypeInfo => {
   return [
     ...cmpMeta.properties.map((cmpProp) => ({