+++ src/mock-doc/css-style-declaration.ts
@@ -88,7 +88,7 @@ function cssCaseToJsCase(str: string) {
       .split('-')
       .map((segment) => segment.charAt(0).toUpperCase() + segment.slice(1))
       .join('');
-    str = str.substr(0, 1).toLowerCase() + str.substr(1);
+    str = str.slice(0, 1).toLowerCase() + str.slice(1);
   }
   return str;
 }