+++ src/sys/node/node-sys.ts
@@ -553,7 +553,7 @@ export function createNodeSys(c: { process?: any } = {}) {
     generateContentHash(content, length) {
       let hash = createHash('sha1').update(content).digest('hex').toLowerCase();
       if (typeof length === 'number') {
-        hash = hash.substr(0, length);
+        hash = hash.slice(0, length);
       }
       return Promise.resolve(hash);
     },
@@ -566,7 +566,7 @@ export function createNodeSys(c: { process?: any } = {}) {
           .on('end', () => {
             let hash = h.digest('hex').toLowerCase();
             if (typeof length === 'number') {
-              hash = hash.substr(0, length);
+              hash = hash.slice(0, length);
             }
             resolve(hash);
           });